pyproject.toml
[project]
name = "code_genereator"
version = "1.0.0"
description = "Agentic Software Factory - A CrewAI Flow-based code generation system"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.80.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
# Main entry points
code_genereator = "code_genereator.main:main"
run_factory = "code_genereator.main:main"
run_flow = "code_genereator.main:run"
kickoff = "code_genereator.main:kickoff"
plot = "code_genereator.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
name = "code_genereator"
